+++ src/components/patients/PatientsList.module.css
@@ -1,0 +1,25 @@
+.patients_list_container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+  width: 100%;
+  padding: 2rem;
+  min-height: 100vh;
+  gap: 1.5rem;
+}
+
+.section_heading {
+  font-size: 1.3rem;
+  color: #025252;
+}
+
+.patients_list {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  align-items: center;
+  justify-content: center;
+  gap: 2rem;
+  height: 100%;
+  list-style: none;
+}
